+++ src/view/pages/server/databases/Databases.vue
@@ -46,10 +46,10 @@
           </b-link>
         </template>
         <template #cell(collation)="data">
-          <i class="rc rc-ln-database rc-table-icon"></i>{{ data.item.collation }}
+          {{ data.item.collation }}
         </template>
         <template #cell(delete)="data">
-          <a role="button" v-on:click="delete_database(data.item._id)"
+          <a role="button" v-on:click="delete_database(data.item)"
             ><i class="fas fa-trash-alt text-danger"></i
           ></a>
         </template>
@@ -62,6 +62,7 @@
 
 <script>
 import { mapGetters } from "vuex";
+import Swal from "sweetalert2";
 import {
   GET_DATABASES,
   DELETE_DATABASE,
@@ -96,12 +97,35 @@ export default {
       });
       this.$router.go();
     },
-    delete_database: function(dbId) {
-      this.$store.dispatch(DELETE_DATABASE, {
-        serverId: this.serverId,
-        databaseId: dbId
+    delete_database: async function(database) {
+      console.log("delete_database", database);
+      const result = await Swal.fire({
+        title: "",
+        text: "Do you want to delete this?",
+        icon: "question",
+        showConfirmButton: true,
+        showCancelButton: true,
+        heightAuto: false
       });
-      this.$store.dispatch(GET_DATABASES, this.serverId);
+      if (!result.isConfirmed) {
+        return;
+      }
+
+      const respone = await this.$store.dispatch(DELETE_DATABASE, {
+        serverId: this.serverId,
+        dbId: database.id
+      });
+      console.log("delete-respone", respone);
+      if (respone && respone.success) {
+        this.$store.dispatch(GET_DATABASES, this.serverId);
+
+        Swal.fire({
+          title: "",
+          text: "Database " + database.name + " has been successfully deleted",
+          icon: "success",
+          heightAuto: false
+        });
+      }
     }
   }
 };
